+++ packages/search/src/types/StoreProps.ts
@@ -11,10 +11,12 @@ import { SpecialZoomLevel } from '@react-pdf-viewer/core';
 import type { MatchPosition } from './MatchPosition';
 import type { NormalizedKeyword } from './NormalizedKeyword';
 import type { SearchTargetPageFilter } from './SearchTargetPage';
+import type { SingleKeyword } from './SingleKeyword';
 
 export interface StoreProps {
     areShortcutsPressed?: boolean;
     doc?: PdfJs.PdfDocument;
+    initialKeyword?: SingleKeyword[];
     keyword?: NormalizedKeyword[];
     matchPosition: MatchPosition;
     renderStatus: Map<number, PluginOnTextLayerRender>;